requirements_viewer.txt
PySide6>=6.6
Pillow>=10.0
UnityPy>=1.20
